# frontend/src/console/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Project",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/signup": {
      "post": {
        "tags": [
          "AUTH ENDPOINTS"
        ],
        "summary": "Signup",
        "operationId": "signup_api_v1_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSignUp"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper_Token_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "AUTH ENDPOINTS"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper_Token_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/users/me": {
      "get": {
        "tags": [
          "AUTH ENDPOINTS"
        ],
        "summary": "Read Users Me",
        "operationId": "read_users_me_api_v1_auth_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper_Dict_str__Any__"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ResponseWrapper_Dict_str__Any__": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ERROR",
              "WARN",
              "SUCCESS"
            ],
            "title": "Status",
            "description": "Status of the response"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Descriptive message about the response"
          },
          "data": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload containing requested data"
          },
          "error": {
            "anyOf": [
              {

              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error details if the status is 'ERROR'"
          },
          "execution_id": {
            "type": "string",
            "title": "Execution Id",
            "description": "Unique ID for tracking the execution"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "ResponseWrapper[Dict[str, Any]]",
        "example": {
          "data": {
            "example_field": "example_value"
          },
          "execution_id": "9f1b6e5d8a7c4b6f8d93e2a1f3a1a9e7",
          "message": "Request processed successfully",
          "status": "SUCCESS"
        }
      },
      "ResponseWrapper_Token_": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ERROR",
              "WARN",
              "SUCCESS"
            ],
            "title": "Status",
            "description": "Status of the response"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Descriptive message about the response"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Token"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload containing requested data"
          },
          "error": {
            "anyOf": [
              {

              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error details if the status is 'ERROR'"
          },
          "execution_id": {
            "type": "string",
            "title": "Execution Id",
            "description": "Unique ID for tracking the execution"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "ResponseWrapper[Token]",
        "example": {
          "data": {
            "example_field": "example_value"
          },
          "execution_id": "9f1b6e5d8a7c4b6f8d93e2a1f3a1a9e7",
          "message": "Request processed successfully",
          "status": "SUCCESS"
        }
      },
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "Access token for authenticated requests"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "Type of token, typically 'Bearer'"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type"
        ],
        "title": "Token",
        "example": {
          "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
          "token_type": "Bearer"
        }
      },
      "UserLogin": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "User's email address"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "User's password for login"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserLogin",
        "example": {
          "email": "user@example.com",
          "password": "securepassword123"
        }
      },
      "UserSignUp": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "User's email address"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "User's password for account creation"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name",
            "description": "User's first name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name",
            "description": "User's last name"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserSignUp",
        "example": {
          "email": "user@example.com",
          "first_name": "John",
          "last_name": "Doe",
          "password": "securepassword123"
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}